+++ scripts/live_test/sendemail.py
@@ -91,7 +91,7 @@ def upload_files(container):
     print('Enter upload_files()')
 
     # Create container
-    cmd = 'az storage container create -n {} --account-name clitestresultstac --account-key {}'
+    cmd = 'az storage container create -n {} --account-name clitestresultstac --account-key {} --public-access container'
     os.popen(cmd.format(container, ACCOUNT_KEY))
 
     # Upload files
@@ -112,20 +112,24 @@ def write_db(container, testdata):
     Insert data to database.
     Sql statements to create table:
     USE clidb;
-    CREATE TABLE t1 (
-       repo       VARCHAR(200) COMMENT 'Repo URL',
-       branch     VARCHAR(200) COMMENT 'Branch name',
-       commit     VARCHAR(50) COMMENT 'Commit ID',
-       target     VARCHAR(2000) COMMENT 'Target modules to test. Splited by space, Empty string represents all modules',
-       live       TINYINT(1) COMMENT 'Live run or not',
-       user       VARCHAR(50) COMMENT 'User (email address) who triggers the run',
-       pass       INT COMMENT 'Number of passed tests',
-       fail       INT COMMENT 'Number of failed tests',
-       rate       VARCHAR(50) COMMENT 'Pass rate',
-       detail     VARCHAR(10000) COMMENT 'Detail',
-       container  VARCHAR(200) COMMENT 'Container URL',
-       date       VARCHAR(10) COMMENT 'Date. E.g. 20200801',
-       time       VARCHAR(10) COMMENT 'Time. E.g. 183000'
+    CREATE TABLE `t1` (
+      `id` int(11) NOT NULL AUTO_INCREMENT,
+      `repr` varchar(30) DEFAULT NULL COMMENT 'date_time_random6digits',
+      `repo` varchar(200) DEFAULT NULL COMMENT 'Repo URL',
+      `branch` varchar(200) DEFAULT NULL COMMENT 'Branch name',
+      `commit` varchar(50) DEFAULT NULL COMMENT 'Commit ID',
+      `target` varchar(2000) DEFAULT NULL COMMENT 'Target modules to test. Splited by space, Empty string represents all modules',
+      `live` tinyint(1) DEFAULT NULL COMMENT 'Live run or not',
+      `user` varchar(50) DEFAULT NULL COMMENT 'User (email address) who triggers the run',
+      `pass` int(11) DEFAULT NULL COMMENT 'Number of passed tests',
+      `fail` int(11) DEFAULT NULL COMMENT 'Number of failed tests',
+      `rate` varchar(50) DEFAULT NULL COMMENT 'Pass rate',
+      `detail` varchar(10000) DEFAULT NULL COMMENT 'Detail',
+      `container` varchar(200) DEFAULT NULL COMMENT 'Container URL',
+      `date` varchar(10) DEFAULT NULL COMMENT 'Date. E.g. 20200801',
+      `time` varchar(10) DEFAULT NULL COMMENT 'Time. E.g. 183000',
+      PRIMARY KEY (`id`),
+      UNIQUE KEY `repr` (`repr`)
     );
 
     """
@@ -139,7 +143,8 @@ def write_db(container, testdata):
                                   host='clisqldbserver.mysql.database.azure.com',
                                   database='clidb')
     cursor = cnx.cursor()
-    sql = 'INSERT INTO t1 (repo, branch, commit, target, live, user, pass, fail, rate, detail, container, date, time) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);'
+    sql = 'INSERT INTO t1 (repr, repo, branch, commit, target, live, user, pass, fail, rate, detail, container, date, time) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);'
+    repr = container
     repo = USER_REPO
     branch = USER_BRANCH
     commit = COMMIT_ID
@@ -154,7 +159,7 @@ def write_db(container, testdata):
     d = datetime.datetime.now()
     date = d.strftime('%Y%m%d')
     time = d.strftime('%H%M%S')
-    data = (repo, branch, commit, target, live, user, pass0, fail, rate, detail, container, date, time)
+    data = (repr, repo, branch, commit, target, live, user, pass0, fail, rate, detail, container, date, time)
     print(data)
     cursor.execute(sql, data)